refactor(protocol): rename stream-utils to async-iterable-utils
Rename to reflect that these are generic async iterable combinators,
not stream-specific. Add JSDoc error-handling docs to each function
and fix split() swallowing unhandled rejections.

Committed-By-Agent: claude
