+++ packages/protocol/src/index.ts
@@ -27,4 +27,4 @@ export {
   drain,
 } from './helpers.js'
 export { parseNdjsonChunks, writeLine } from './ndjson.js'
-export { channel, merge, split, map } from './stream-utils.js'
+export { channel, merge, split, map } from './async-iterable-utils.js'
